v1.0.1 - Cleaned header files, added file size limit for the sake of memory efficiency

+++ src/global/global.c
@@ -2,7 +2,7 @@
 #include <stdlib.h>
 
 int length = 0;
-unsigned long fileSize = 0;
+int fileSize = 0;
 FILE *input;
 FILE *output;
